+++ API/DTOs/TaxReturnDto.cs
@@ -4,7 +4,7 @@
     {
         public Guid Id { get; set; }
         public string TaxName { get; set; }
-        public int TaxRate { get; set; }
+        public decimal TaxRate { get; set; }
         public bool CompoundTax { get; set; }
     }
 }
